+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.13.0-rc.21",
-  genVersion: "2.708.2",
-  userAgent: "speakeasy-sdk/typescript 3.13.0-rc.21 2.708.2 2.0 @orq-ai/node",
+  sdkVersion: "3.13.0-rc.22",
+  genVersion: "2.709.0",
+  userAgent: "speakeasy-sdk/typescript 3.13.0-rc.22 2.709.0 2.0 @orq-ai/node",
 } as const;