+++ packages/orq-rc/RELEASES.md
@@ -3749,3 +3749,13 @@ Based on:
 - [typescript v3.13.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.21 - packages/orq-rc
+
+## 2025-09-22 04:42:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.622.1 (2.709.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.22 - packages/orq-rc